chore(ui): move deprecated components to deprecated_js
