+++ packages/ui-components/src/deprecated_js/AppIntro/AppIntro.component.tsx
