+++ packages/ui-components/src/deprecated_js/AppIntro/AppIntro.test.tsx
