+++ packages/ui-components/src/deprecated_js/FilterInput/FilterInput.component.js
@@ -5,10 +5,10 @@
 
 import React, { useState, useEffect } from "react"
 import PropTypes from "prop-types"
-import { NativeSelect } from "../../deprecated_js/NativeSelect/NativeSelect.component"
-import { NativeSelectOption } from "../../deprecated_js/NativeSelectOption/NativeSelectOption.component"
-import { TextInput } from "../../deprecated_js/TextInput/TextInput.component"
-import { Icon } from "../../deprecated_js/Icon/Icon.component"
+import { NativeSelect } from "../NativeSelect/NativeSelect.component"
+import { NativeSelectOption } from "../NativeSelectOption/NativeSelectOption.component"
+import { TextInput } from "../TextInput/TextInput.component"
+import { Icon } from "../Icon/Icon.component"
 
 const wrapperStyles = `
 	jn-flex
