+++ packages/ui-components/src/deprecated_js/FilterInput/index.js
