+++ packages/ui-components/src/deprecated_js/FilterPill/FilterPill.stories.js
