+++ packages/ui-components/src/deprecated_js/SwitchRow/SwitchRow.test.js
