+++ packages/ui-components/src/deprecated_js/TextareaRow/TextareaRow.component.js
@@ -5,9 +5,9 @@
 
 import React from "react"
 import PropTypes from "prop-types"
-import { Textarea } from "../../deprecated_js/Textarea/index.js"
-import { FormRow } from "../../deprecated_js/FormRow/FormRow.component"
-import { withDeprecationWarning } from "../../deprecated_js/withDeprecationWarning/index.js"
+import { Textarea } from "../Textarea/index.js"
+import { FormRow } from "../FormRow/FormRow.component.js"
+import { withDeprecationWarning } from "../withDeprecationWarning/index.js"
 
 /** DEPRECATED: A textarea row containing a textarea, associated label, optional helptext, and structural markup. This component is DEPRECATED, use Textarea instead. */
